+++ CHANGELOG.md
@@ -1,11 +1,24 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
+Release history for pitchfork. For installation and current usage, see the
+[documentation](https://pitchfork.jdx.dev).
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.24.2](https://github.com/jdx/pitchfork/compare/v2.24.1...v2.24.2) - 2026-09-05
+
+### Other
+
+- publish a patch release with the current packaged contents
+
+## [2.24.1](https://github.com/jdx/pitchfork/compare/v2.24.0...v2.24.1) - 2026-09-02
+
+### Other
+
+- *(deps)* update rust crate usage_rs to v6.4.1 ([#804](https://github.com/jdx/pitchfork/pull/804))
 
 ## [2.24.0](https://github.com/jdx/pitchfork/compare/v2.23.0...v2.24.0) - 2026-08-31